Fixing eslint errors in test utils.ts
