+++ packages/teams-js/src/public/profile.ts
@@ -1,5 +1,6 @@
 import { sendAndHandleSdkError as sendAndHandleError } from '../internal/communication';
 import { ensureInitialized } from '../internal/internalAPIs';
+import * as internalProfile from '../internal/profile';
 import { validateShowProfileRequest } from '../internal/profileUtil';
 import { FrameContexts } from './constants';
 import { ErrorCode } from './interfaces';
@@ -11,6 +12,14 @@ import { runtime } from './runtime';
  * @beta
  */
 export namespace profile {
+  export import Modality = internalProfile.Modality;
+  export import Persona = internalProfile.Persona;
+  // Even though this type is unused in this file, it is referenced by Persona and thus should
+  // be re-exported to ensure it can be used and be documented publicly.
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  export import PersonaIdentifiers = internalProfile.PersonaIdentifiers;
+  export import TriggerType = internalProfile.TriggerType;
+
   /**
    * Opens a profile card at a specified position to show profile information about a persona.
    * @param showProfileRequest The parameters to position the card and identify the target user.
@@ -28,7 +37,7 @@ export namespace profile {
       }
 
       // Convert the app provided parameters to the form suitable for postMessage.
-      const requestInternal: ShowProfileRequestInternal = {
+      const requestInternal: internalProfile.ShowProfileRequestInternal = {
         modality: showProfileRequest.modality,
         persona: showProfileRequest.persona,
         triggerType: showProfileRequest.triggerType,
@@ -42,72 +51,6 @@ export namespace profile {
 
       resolve(sendAndHandleError('profile.showProfile', requestInternal));
     });
-  }
-
-  /**
-   * The type of modalities that are supported when showing a profile.
-   * Can be provided as an optional hint with the request and will be
-   * respected if the hosting M365 application supports it.
-   *
-   * @beta
-   */
-  export type Modality = 'Card' | 'Expanded';
-
-  /**
-   * The type of the profile trigger.
-   *  - MouseHover: The user hovered a target.
-   *  - Press: The target was pressed with either a mouse click or keyboard key press.
-   *  - AppRequest: The show profile request is happening programmatically, without direct user interaction.
-   *
-   * @beta
-   */
-  export type TriggerType = 'MouseHover' | 'Press' | 'AppRequest';
-
-  /**
-   * The set of identifiers that are supported for resolving the persona.
-   *
-   * At least one is required, and if multiple are provided then only the highest
-   * priority one will be used (AadObjectId > Upn > Smtp).
-   *
-   * @beta
-   */
-  export type PersonaIdentifiers = {
-    /**
-     * The object id in Azure Active Directory.
-     *
-     * This id is guaranteed to be unique for an object within a tenant,
-     * and so if provided will lead to a more performant lookup. It can
-     * be resolved via MS Graph (see https://learn.microsoft.com/graph/api/resources/users
-     * for examples).
-     */
-    readonly AadObjectId?: string;
-
-    /**
-     * The primary SMTP address.
-     */
-    readonly Smtp?: string;
-
-    /**
-     * The user principle name.
-     */
-    readonly Upn?: string;
-  };
-
-  /**
-   * The persona to show the profile for.
-   *
-   * @beta
-   */
-  export interface Persona {
-    /**
-     * The set of identifiers that are supported for resolving the persona.
-     */
-    identifiers: PersonaIdentifiers;
-
-    /**
-     * Optional display name override. If not specified the user's display name will be resolved normally.
-     */
-    displayName?: string;
   }
 
   /**
@@ -134,27 +77,6 @@ export namespace profile {
     /**
      * Specifies which user interaction was used to trigger the API call.
      */
-    triggerType: TriggerType;
-  }
-
-  /**
-   * Internal representation of a DOMRect suitable for sending via postMessage.
-   */
-  type Rectangle = {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  };
-
-  /**
-   * An internal representation of the showProfile parameters suitable for sending via postMessage.
-   * The hub expects to receive an object of this type.
-   */
-  interface ShowProfileRequestInternal {
-    modality?: Modality;
-    persona: Persona;
-    targetRectangle: Rectangle;
     triggerType: TriggerType;
   }
